requirements.txt
flask
pdfplumber
python-docx
google-generativeai
fpdf
werkzeug
python-dotenv==1.0.0